added glow for your own team

+++ MultiHack/class/Memory/memory.cpp
@@ -45,16 +45,6 @@ DWORD Memory::getModule(char *ModuleName){
 
 int Memory::read(DWORD address){
     DWORD value;
-    ReadProcessMemory(hProcess, (LPVOID)address, &value,sizeof(long),0);
+    ReadProcessMemory(hProcess, (LPVOID)address, &value,sizeof(DWORD),0);
     return value;
 }
-
-
-
-//void Memory::write(DWORD dwAddress, DWORD value){
-//    WriteProcessMemory(hProcess, (LPVOID)dwAddress, &value, sizeof(long), 0);
-//}
-
-//void Memory::writeFloat(DWORD dwAddress, float value){
-//    WriteProcessMemory(hProcess, (LPVOID)dwAddress, &value, sizeof(float), NULL);
-//}
